requirements.txt
fastapi
uvicorn
jinja2
python-multipart
boto3
python-dotenv